+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+* `vendorCargoDeps` correctly accepts arguments which have _not_ set `src`, so
+  long as one of `cargoLock`, `cargoLockContents`, or `cargoLockParsed` is set
+
 ## [0.12.0] - 2023-03-19
 
 ### Added